+++ js/editor.js
@@ -149,6 +149,7 @@
                         config.events.onRequestSelectDocument = OCA.Onlyoffice.onRequestSelectDocument;
                         config.events.onRequestSendNotify = OCA.Onlyoffice.onRequestSendNotify;
                         config.events.onRequestReferenceData = OCA.Onlyoffice.onRequestReferenceData;
+                        config.events.onRequestOpen = OCA.Onlyoffice.onRequestOpen;
                         config.events.onMetaChange = OCA.Onlyoffice.onMetaChange;
 
                         if (OC.currentUser) {
@@ -551,6 +552,14 @@
                     timeout: 3
                 });
             });
+    };
+
+    OCA.Onlyoffice.onRequestOpen = function (event) {
+        let filePath  = event.data.path;
+        let fileId = event.data.referenceData.fileKey;
+        let windowName = event.data.windowName;
+        let sourceUrl = OC.generateUrl(`apps/${OCA.Onlyoffice.AppName}/${fileId}?filePath=${OC.encodePath(filePath)}`);
+        window.open(sourceUrl, windowName);
     };
 
     OCA.Onlyoffice.onMetaChange = function (event) {